requirements.txt
# wechat-tech-writer 依赖
requests>=2.28.0
google-genai>=1.0.0
openai>=1.0.0

# wechat-article-formatter 依赖
markdown>=3.4.0
beautifulsoup4>=4.11.0
cssutils>=2.6.0

# wechat-draft-publisher 依赖
requests>=2.28.0
beautifulsoup4>=4.11.0